fix database typo + datetime
