+++ src/cmd/api/models/video.go
@@ -22,21 +22,3 @@ type Video struct {
 	CreatedAt  *time.Time
 	UpdatedAt  *time.Time
 }
-
-type UploadStatus int
-
-const (
-	STARTED UploadStatus = iota
-	DONE
-	FAILED
-)
-
-type Upload struct {
-	ID         string
-	VideoId    string
-	Status     UploadStatus
-	UploadedAt *time.Time
-	CreatedAt  *time.Time
-	UpdatedAt  *time.Time
-	Progress   int
-}
